Use a POLYUNSIGNED for StreamToken.streamNo
On a 64-bit big-endian architecture with sizeof(unsigned) == 4, streamNo
is set for stdin/stdout/stderr in add_word_to_io_area by treating it as
a PolyWord, which is a 64-bit value. However, streamNo is a 32-bit
value, so any code that uses it sees 4 bytes of zeroes, as the
least-significant bits are in the next 4 bytes. By using a POLYUNSIGNED,
the sizes match.