Don't support rounding modes for soft-float systems
On armel, floating point operations are implemented in software.
However, the implementation always rounds to nearest. If the system has
a hardware FPU, it will support setting the rounding mode to any valid
value, and subsequent gets will return this value, but it will have no
effect on the actual rounding performed. If the system doesn't have a
hardware FPU, it will only allow the rounding mode to be set to
FE_NEAREST, and getting the rounding mode will always return this.

Since rounding may not be supported, the floating point rounding mode
test needs to be skipped on unsupported systems. This is implemented by
raising Fail "Skip Test", which is caught in runTests.