libpolyml: Bail out when file operations fail during state loading.
This commit modifies the state loading logic to return when file operations
fail while processing relocations. The rationale is that, while it may make
sense to continue when a relocation entry is corrupted, is does not make sense
to continue when an fread or fseek fails as we end up incorrectly interpreting
data.