Always include long long Make_fixed_precision
One of the users of this function is getStatInfo in unix_specific, which
converts a struct stat's st_nlink. However, on x32, st_nlink is a 64-bit
quantity (the system call ABI is 64-bit), so there are two problems.
Firstly, it will be truncated; secondly, since the long and int
overloads involve a cast to an integer of the same width, compilation
fails with an ambiguous call.